permissions.insync? for Windows provider
This provides a check based on whether permissions are in sync, taking into
account whether the purge option was chosen or not.

This takes the current and should permissions and generates a sync checking
array for each of them that contains only the elements in aces that would
constitute a difference, and ensures the names would match the machine specific
names and SIDs.