(PUP-1737) Patch win32-security for legacy versions
This relates to PUP-1284 with a backwards compatibility to older versions of
puppet that have not been patched. This will allow the ACL provider to work
properly with those older versions by correcting unsafe stripping of the SIDs
when something appaears to be whitespace.

This checks the gem version of win32-security and only applies the patch if the
version is known to have the incorrect functionality.

See https://github.com/djberg96/win32-security/issues/3 for more details.