(PUP-847) Group has no default
From http://msdn.microsoft.com/en-us/library/windows/desktop/ms676927(v=vs.85).aspx
it has been noted that there really isn't a default group on Windows, so trying
to specify this could lead to issues of sync. If someone has not specified a group,
then we shouldn't try to manage it. This commit changes the behavior to allow
group to stay as it is unless explicitly managed.