(PUP-851) Manage propagation (PUP-846) Manage Inheritance
This adds the inheritance and propagation flags for each ace based on how they
are set up in the manifest. Also if someone sets `child_types => 'none'` and
changes `affects` to anything but 'self_only', a warning will be passed back
with a note on how to get rid of it. In cases where the user does this, the
provider will react appropriately and add the warning instead of erroring.