(doc) add more docs to permissions/owner/group
This updates permissions property with quite a bit more information for usage.
For owner and group it just adds that the account must both already exist
and that the provider attempts to autorequire the user for permission
identities, owner and group. This brings to light the fact that we are not also
looking at autorequire of group so a note was set on that.