(PUP-845) exists?, create, destroy (partial)
This provides better handling on exists? and create since the type only manages
the permissions associated with a target and not the creation or removing of
the target itself.

For destroy this returns an error message if permissions are not also
specified. In most cases they would be, except when run from puppet resource.