(PUP-844) Ace - check for id.nil on get
Instead of relying on the set method to properly set the id for an ace, attempt
to use the provider method when getting the id instead. This removes the order
that identity and id will need to be set in and further allows for checking the
id value later when being used with items that have not set the value.