(maint) Remove creation of methods on provider
Prior to this during autorequires, if a method didn't exist on the provider we
would create it with a block to just return the value passed in. This saved a
few repeated checks. This could be confusing during inspection so we favored a
more explicit approach to see if the method exists and then only call it when
it does.

Paired with Josh Cooper <josh@puppetlabs.com>