(PUP-1988) Ensure=>absent logs property changes
This implementation hooks into ResourceHarness to get the current instance and
allow syncing of properties when ensure changes. This code performs much of the
same steps as  ResourceHarness.perform_changes does, but only during destroy()
(ensure => absent). This allows the messages not only to be logged to the
console about changes to properties, but also to end up in reports.

This code will be removed by a better solution.