(PUP-1988) ensure=>absent to purge=>listed_permissions
In providing good output for `ensure=>absent`, we realized that using ensurable
was the incorrect way of providing a way to remove permissions. In doing that
we've decided to use `purge=>listed_permissions` as a better way of making the
manifest easier to understand just by looking at it and to make the code base
more maintainable.