(PUP-1893) Formatting puppet resource
`Puppet resource acl location` does not use the same formatting that the object
uses if you implement Puppet::Parameter.format_value_for_display as a custom
change for your particular parameter/property. This is due to behavior in
Puppet::Resource.to_manifest not having access to the full resource and makes a
limited call on formatting the value with
Puppet::Parameter.format_value_for_display without even a should_to_s or
is_to_is (which could help out immensely if it did).

This makes puppet resource believe that our custom Ace object is actually a
hash so that it will have proper formatting for copying into a manifest.