(PUP-2238) Adjust file permission defaults
A file doesn't have inheritance, so it should be
`affects => 'self_only'` when the target path is a file
and not a directory.  Without this users would need
to ensure they always specifiy `affects => 'self_only'`
for every identity in the acl resource when targeting
a file and not a directory.