(doc) How acl module defines unique ACEs
This is important due to how Windows can technically allow you to have ACEs
that differ only in rights assignment. That is not manageable from the acl
module as it would see an ACL with those ACEs out of sync since they would
match up with more than one ACE.

ACEs are defined to be unique based on the combination of identity, type,
child_types, and affects.