(PUP-2338) fail on symbolic link
Prior to this if a user specified a symlink as the target of the acl, puppet
would not report an error during application. It would just silently fail.
Since the acl module does not support symlinks so the appropriate
behavior is to throw an error informing the user that symlinks are not
supported.