(PUP-1893) Formatting specs
The formatting for when it is called when applying the changes versus when
puppet resource calls it is drastically different. It will be more preferable to
see puppet enhance the use of puppet resource to use the one that each
parameter has set up instead of the default which allows for zero custom
formatting.