(MODULES-1482) Fix Autorequire to only include resource title
Previously the acl module was using autorequires that include the resource name
and type, like Group['None']. With PUP-3177 (first available in Puppet 3.7.1
and in PE 3.7.0), this call now results in invalid calls to `Puppet::Resource`.

Specifying the resource type is incorrect and not a best practice (it
is already handled by `autorequire(:type)` at the start of each of the
autorequire methods.

This adjusts the autorequires to follow the correct practice of only including
the resource title.

This also adjusts file autorequires to greatly simplify the logic and reduce
the extra call to `catalog.resource` which is called later by
`Puppet::Resource`. The reduction in logic means the file resource name case
needs to match exactly, but that is in line with how other requires should be
specified.