(maint) Fix specs
Rspec 3+ requires that should be explicitly set as an expect term.
be_true and be_false have also been changed so that they are not
whether something is truthy or falsey, but simply really true/false. To
get the older behavior, the test simply needs to check for be_truthy or
be_falsey. For explicit true/false checks, it is be true and be false.