(MODULES-4230) Add support for custom puppet-agent installations
Previously the beaker pupet installation helper is only able to use published
puppet agent version.  In order to satisfy PA-562 it should be possible to set
a custom puppet-agent version to be used during the acceptance tests.

This commit takes the puppet agent installation logic from the various
puppet-agent component pipelines and changes the behavior of the puppet-agent
installation based on the presence of the SHA environment variable.  This
variable is set during CI and local developer workflows when using a custom
puppet-agent build.