(maint) Update tests to properly interpret Unicode
Previously, the acceptance tests were accidentally working due to a combination
of ruby, cygwin and beaker all correctly corrupting Unicode characters when
transmitted over STDIN/STDOUT.  This commit changes the communication transports
used to ones that retain the unicode information and are succesfully applied on
the system under test.  This commit also removes the file content tests as they
are better served by using unique test fixture names instead of relying on the
file content to be valid.