(maint) Modify beaker rake task to invoke beaker directly
The acceptance tests for ACL are written in standard beaker format however
the preferred method is using beaker-rspec.  This commit overrides the
default `beaker` task, which would normally use beaker-rspec, and instead
invokes beaker directly.  This is only need while the module tests are migrated
to the newer rspec-beaker format.

[skip ci]