(MODULES-5899) Fix code for rubocop violations
This commit updates the non-test code to repair rubocop violations.  Note that
a rubocop_todo file has been introduced.  These violations should be fixed in
the future but will require more effort than can be expected in this commit.
Later commits will fix the violations and remove the todo entries.